+++ docs/src/js/components/architecture/architecture.jsx
@@ -6,77 +6,55 @@ export default class Architecture extends PureComponent {
   render() {
     return <Container className="main container scrollable">
       <h1 className="mb-md">Application Architecture</h1>
-      <p className="mb-md">
-        Extension React provides support for MVC+VM application architectures. 
-        To understand what is MVC+VM, we should start by further defining what the various abbreviations represent.
-      </p>
-      <ul className="mb-md">
-        <li>
-          <strong>(M) Model</strong> - This is the data for your application.
-          A set of classes (called "Models") defines the fields for their data (e.g. a User model with user-name and password fields).
-          Models know how to persist themselves through the data package and can be linked to other models via associations.
-          Models are normally used in conjunction with Stores to provide data for grids and other components.
-          Models are also an ideal location for any data logic that you may need, such as validation, conversion, etc.
-        </li>
-        <li>
-          <strong>(V) View</strong> - A View is any type of component that is visually represented. For instance, grids, trees and panels are all considered Views.
-        </li>
-        <li>
-          <strong>(C) Controller</strong> - Controllers are used as a place to maintain the view's logic that makes your app work.
-          This could entail rendering views, routing, instantiating Models, and any other sort of app logic.
-        </li>
-        <li>
-          <strong>(VM) ViewModel</strong> - The ViewModel is a class that manages data specific to the View.
-          It allows interested components to bind to it and be updated whenever this data changes.
-        </li>
-      </ul>
-      <h3 className="mb-md">Models and Stores</h3>
-      <p className="mb-md">
-        <code>Models</code> and <code>Stores</code> make up the information gateway of your application.
-        Most of your data is sent, retrieved, organized, and "modeled" by these two classes.
-      </p>
-      <p className="mb-md">
-        A <code>Model</code> represents any type of persist-able data in your application.
-        Each model has fields and functions that allow your application to "model" data.
-        Models are most commonly used in conjunction with stores.
-        Stores can then be consumed by data-bound components like grids, trees, and charts.
-      </p>
-      <p className="mb-md">
-        A <code>Store</code> is a client side cache of records (instances of a Model class).
-        Stores provide functions for querying the records contained within.
-      </p>
-      <h3 className="mb-md">ViewControllers</h3>
-      <p className="mb-md">Extension React delivers some exciting improvements for React. To enhance MVC applications, Extension React provides ViewControllers, also called as Component:</p>
-      <ul className="mb-md">
-        <li>Simplifies the connection to views using “Component” decorator.</li>
-        <li>Leverages the life cycle of views to automatically manage their associated.</li>
-        <li>Reduces complexity based on a one-to-one relationship with the managed view.</li>
-        <li>Provides encapsulation to make nesting views reliable.</li>
-        <li>Retains the ability to select components and listen to their events at any level below the associated view.</li>
-      </ul>
-      <pre>
+      <pre className="mb-md">
+{`// ./app.js
+import 'babel-polyfill';
+import React from 'react';
+import Rext from 'ext-react';
+import Viewport from './components/viewport/viewport';
+
+Rext.application({
+  stores: [
+    require('./stores/data'),
+  ],
+  views: [
+    require('./components/search/search'),
+  ],
+  launch: () => <Viewport />
+});`}
+      </pre>
+      <pre className="mb-md">
+{`// ./stores/card
+import { Store } from 'ext-react';
+
+export default Store({
+  storeId: 'DataStore',
+  proxy: {
+    url: '/data/sample.json'
+  }
+})`}
+      </pre>
+      <pre className="mb-md">
 {`// ./components/search/search.js
-import React, { Component } from 'react';
+import React from 'react';
 import Rext, { Container } from 'ext-react';
-import DataStore from '~/stores/data';
 import SearchForm from './search-form';
 import SearchResult from './search-result';
 
-export default class Search extends Component {
-  render() {
-    return <Container>
-      <SearchForm />
-      <SearchResult />
-    </Container>
-  }
-}
-
-// ./components/search/search-form.js
-import { Component } from 'ext-react';
-import DataStore from '~/stores/data';
+export function Search() {
+  return <Container>
+    <SearchForm />
+    <SearchResult />
+  </Container>
+}`}
+      </pre>
+      <pre className="mb-md">
+{`// ./components/search/search-form.js
+import { Component, bind } from 'ext-react';
 import SearchFormView from './search-form.view';
 
 @Component({
+  store: [ 'DataStore' ]
   view: SearchFormView
 })
 export default class SearchForm {
@@ -90,8 +68,9 @@ export default class SearchForm {
     };
   }
 
+  @bind
   search(criteria) {
-    criteria = this.correctCriteria(criteria);
+    const { DataStore } = this.stores;
     DataStore.rejectChanges();
     Rext.extend(DataStore.proxy, {
       params: criteria,
@@ -103,18 +82,21 @@ export default class SearchForm {
     DataStore.load();
   }
 
+  @bind
   clearSearchResult(comp) {
+    const { DataStore } = this.stores;
     DataStore.rejectChanges();
     DataStore.clearData();
     comp.setState(() => (this.criteria));
   }
-}
-
-// ./components/search/search-form.view.jsx
-import React, { Component } from 'react';
+}`}
+      </pre>
+      <pre className="mb-md">
+{`// ./components/search/search-form.view.jsx
+import React, { PureComponent } from 'react';
 import { Field, Dropdown, Button } from 'ext-react';
 
-export default class SearchFormView extends Component {
+export default class SearchFormView extends PureComponent {
   constructor(props) {
     super(props);
     Ext.initialState(this, props.$view.criteria);
@@ -133,19 +115,23 @@ export default class SearchFormView extends Component {
       <Button text="Clear" onClick={() => clearSearchResult(this)} />
     </section>
   }
-}
+}`}
+      </pre>
+      <pre className="mb-md">
+{`// ./components/search/search-result.jsx
+import React, { PureComponent } from 'react';
+import Rext, { Container, Grid } from 'ext-react';
 
-// ./components/search/search-result.jsx
-import React, { Component } from 'react';
-import { Container, Grid } from 'ext-react';
-import DataStore from '~/stores/data';
+export default class SearchResult extends PureComponent {
+  constructor() {
+    this.DataStore = Rext.StoreManager.get('DataStore');
+  }
 
-export default class SearchResult extends Component {
   render() {
     const { name, recordStatuses, purposes, activities, products } = this.state,
           { search, clearSearchResult } = this.props.$view;
     return <Container>
-      <Grid store={DataStore}>
+      <Grid store={this.DataStore}>
         <div text="ID" dataIndex="id" />
         <div text="Name" dataIndex="name" />
         <div text="Status" dataIndex="status" />