+++ docs/src/js/components/example/cards.js
@@ -1,19 +1,18 @@
 import React, { PureComponent } from 'react';
 import { Route, Component } from '~/rext';
-import CardStore from './store/card';
 
 class CardView extends PureComponent {
   componentDidMount() {
-    CardStore.load();
+    this.props.stores.CardStore.load();
   }
-  render() {console.log('render card view')
+  render() {
     return <section />
   }
 }
 
 @Route('/example/cards')
 @Component({
-  stores: [ CardStore ],
+  stores: [ 'CardStore' ],
   view: CardView
 })
 export default class Card { }